case1 scene1,2,3,4
This code measures the position of obstacles using a LIDAR, where the angle of 0 represents the height of the court, and the angle of 90 represents the width of the court. Thus, the court is divided into four quadrants, and if the obstacle is located in scene1, 2, 3, or 4, the robot moves along the path according to that scene. While the height and width of the court can be calculated relatively accurately using LIDAR, the measurement of the x and y coordinates of the obstacle failed due to values that spike like noise.