distance maintenance driving
Before starting to drive, the initial distance is set by measuring the distance to the side with a LiDAR, and then the drive begins while maintaining this initial distance. If an obstacle or wall appears ahead, you must make a U-turn unconditionally. The first U-turn is to the right, and the second U-turn is to the left. From then on, U-turns are made alternately.